pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "tt-adapter"
version = "0.0.1"
description = "Model Explorer Adapter built for TT-MLIR Compiled outputs."
readme = "README.md"
dependencies = [
    "ai-edge-model-explorer>=0.1"
]